Added fcm360w_driver, update drivers structure, can0_driver update

+++ PIC32CK_DRIVER/src/config/default/system/time/src/sys_time_local.h
@@ -111,7 +111,7 @@ typedef struct{
     SYS_STATUS status;
     const SYS_TIME_PLIB_INTERFACE*  timePlib;
     INT_SOURCE                      hwTimerIntNum;
-    float                           hwTimerTickFreq;
+    uint32_t                        hwTimerTickFreq;
     uint32_t                        hwTimerPeriodValue;
     volatile uint64_t               swCounter64;           /* Software 64-bit counter */
     uint8_t                         interruptNestingCount;
